requirements.txt
gradio
langchain
langchain-openai
langchain-community
faiss-cpu
huggingface_hub
transformers
sentence-transformers
pypdf
python-dotenv
fpdf
